update config variables
