+++ packages/cardpay-reward-programs/cardpay_reward_programs/utils.py
@@ -1,3 +1,4 @@
+import json
 from os import stat
 from pathlib import PosixPath
 
@@ -70,3 +71,32 @@ def get_partition_files(config_location, table, min_partition, max_partition):
 def get_files(config_location, table, min_partition, max_partition):
     file_list = get_partition_files(AnyPath(config_location), table, min_partition, max_partition)
     return list(map(get_local_file, file_list))
+
+
+def select_rule(name, core_parameters, user_defined_parameters):
+    assert name
+    rule_constructor = globals()[name]
+    instance = rule_constructor(core_parameters, user_defined_parameters)
+    return instance
+
+
+def blob_to_rule(parameters):
+    core_parameters, user_defined_parameters = get_parameters(parameters)
+    name = to_camel_case(core_parameters["docker_image"])
+    return select_rule(name, core_parameters, user_defined_parameters)
+
+
+def get_parameters(parameters):
+    """
+    TODO: take hex blob as input instead of parameters
+    """
+    core_parameters = parameters.get("core")
+    user_defined_parameters = parameters.get("user_defined")
+    return core_parameters, user_defined_parameters
+
+
+def to_camel_case(snake_str):
+    components = snake_str.split("_")
+    # We capitalize the first letter of each component except the first one
+    # with the 'title' method and join them together.
+    return components[0] + "".join(x.title() for x in components[1:])
